refactor and add tests
